+++ src/all.ts
@@ -1,5 +1,6 @@
 // Components
 export { CpsAccordion } from './components/accordion.js';
+export { CpsAccordionGroup } from './components/accordion-group.js';
 export { CpsAvatar } from './components/avatar.js';
 export { CpsBackground } from './components/background.js';
 export { CpsBadge } from './components/badge.js';